Fix BNCI2015_006 and clean up legacy imports
- Fix BNCI2015_006 subject codes and URL format for download
- Fix trial indices type error in data loading
- Update summary_p300.csv with actual trial counts
- Remove deprecated aliases from datasets __init__.py
- Rename legacy_base.py to base.py and merge with existing base
- Add description consistency fix for MNE Raw concatenation
