+++ moabb/datasets/summary_p300.csv
@@ -31,6 +31,6 @@ BNCI2015_012, 10, 63, 6075 NT / 759 T, 1.0, 1000, 1,
 BNCI2015_013, 6, 64, 809 NT / 235 T, 0.6, 512, 2,
 BNCI2016_002, 15, 69, varies brake / EMG, 1.5, 200, 1,
 BNCI2020_002, 18, 31, varies NT / T, 16, 250, 1,
-BNCI2015_006, 11, 64, varies attended / unattended, 1.0, 100, 1,
+BNCI2015_006, 11, 64, ~875 unattended / ~856 attended, 1.0, 100, 1,
 BNCI2015_007, 16, 63, varies NT / T, 0.7, 100, 1,
 BNCI2015_008, 13, 63, varies NT / T, 1.0, 250, 2,
